@@ -35,7 +35,7 @@ class _PhoneVerificationScreenState extends State<PhoneVerificationScreen> {
     return Scaffold(
       appBar: PreferredSize(preferredSize: const Size.fromHeight(kToolbarHeight), child: AppBarWidget(title: PhoneVerificationScreenViewModel().title)),
       body: Padding(
-        padding: const EdgeInsets.all(mainUnit),
+        padding: EdgeInsets.all(Units().spacing),
         child: Column(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: [
@@ -45,7 +45,7 @@ class _PhoneVerificationScreenState extends State<PhoneVerificationScreen> {
               child: Column(
                 children: [
                   PINCodeInputWidget(numbers: _numbers),
-                  const SizedBox(height: mainUnit),
+                  SizedBox(height: Units().spacing),
                   NumberPadWidget(numbers: _numbers, update: _update),
                 ],
               ),
