@@ -10,17 +10,17 @@ class Setting extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      padding: const EdgeInsets.symmetric(horizontal: mainUnit / 2, vertical: mainUnit - 8),
-      margin: const EdgeInsets.only(bottom: mainUnit / 2),
-      decoration: BoxDecoration(color: colorOnBackground, borderRadius: BorderRadius.circular(borderRadius)),
+      padding: EdgeInsets.symmetric(horizontal: Units().spacing / 2, vertical: Units().spacing - 8),
+      margin: EdgeInsets.only(bottom: Units().spacing / 2),
+      decoration: BoxDecoration(color: Theme.of(context).colorScheme.onBackground, borderRadius: BorderRadius.circular(Units().borderRadius)),
       clipBehavior: Clip.hardEdge,
       child: Row(
         mainAxisAlignment: MainAxisAlignment.start,
         crossAxisAlignment: CrossAxisAlignment.center,
         children: [
-          Icon(icon, color: colorTertiary, size: 24),
-          const SizedBox(width: mainUnit / 2),
-          Text(title, style: Theme.of(context).textTheme.bodyMedium?.copyWith(color: colorTertiary)),
+          Icon(icon, color: Theme.of(context).colorScheme.tertiary, size: 24),
+          SizedBox(width: Units().spacing / 2),
+          Text(title, style: Theme.of(context).textTheme.bodyMedium?.copyWith(color: Theme.of(context).colorScheme.tertiary)),
         ],
       ),
     );
