@@ -6,19 +6,19 @@ void showModal(dynamic context, String title, Widget body) {
     context: context,
     isScrollControlled: true,
     clipBehavior: Clip.hardEdge,
-    barrierColor: colorTertiary.withOpacity(0.25),
-    shape: const RoundedRectangleBorder(
+    barrierColor: Theme.of(context).colorScheme.tertiary.withOpacity(0.25),
+    shape: RoundedRectangleBorder(
       borderRadius: BorderRadius.only(
-        topLeft: Radius.circular(borderRadius),
-        topRight: Radius.circular(borderRadius),
+        topLeft: Radius.circular(Units().borderRadius),
+        topRight: Radius.circular(Units().borderRadius),
       ),
     ),
-    backgroundColor: colorBackground,
+    backgroundColor: Theme.of(context).colorScheme.background,
     builder: (BuildContext context) {
       return Padding(
-        padding: const EdgeInsets.all(mainUnit),
+        padding: EdgeInsets.all(Units().spacing),
         child: Wrap(
-          runSpacing: mainUnit,
+          runSpacing: Units().spacing,
           alignment: WrapAlignment.center,
           children: [
             Text(title, style: Theme.of(context).textTheme.titleMedium),
