@@ -1,13 +1,28 @@
 import 'package:flutter/material.dart';
 
-const Color colorPrimary = Color(0xFFFFFFFF);
-const Color colorSecondary = Color(0xFFEEEEEE);
-const Color colorTertiary = Color(0xFFAAAAAA);
-const Color colorBackground = Color(0xFF000000);
-const Color colorOnBackground = Color(0xFF151515);
-const Color colorGreen = Color(0xFF54B435);
-const Color colorRed = Color(0xFFF32424);
-const Color colorGold = Color(0xFFFFD700);
+class LightThemeColors {
+  final Color primary = const Color(0xFF000000);
+  final Color secondary = const Color(0xFF151515);
+  final Color tertiary = const Color(0xFFAAAAAA);
+  final Color background = const Color(0xFFFFFFFF);
+  final Color onBackground = const Color(0xFFEEEEEE);
+}
 
-const double mainUnit = 24.0;
-const double borderRadius = 12.0;
+class DarkThemeColors {
+  final Color primary = const Color(0xFFFFFFFF);
+  final Color secondary = const Color(0xFFEEEEEE);
+  final Color tertiary = const Color(0xFFAAAAAA);
+  final Color background = const Color(0xFF000000);
+  final Color onBackground = const Color(0xFF151515);
+}
+
+class GlobalColors {
+  final Color success = const Color(0xFF54B435);
+  final Color error = const Color(0xFFF32424);
+  final Color pendding = const Color(0xFFFFD700);
+}
+
+class Units {
+  final double spacing = 24.0;
+  final double borderRadius = 12.0;
+}
