@@ -1,4 +1,3 @@
-import 'package:lexa/view_model/theme/constants.dart';
 import 'package:flutter/material.dart';
 
 class TextInputWidget extends StatelessWidget {
@@ -29,7 +28,7 @@ class TextInputWidget extends StatelessWidget {
       textInputAction: TextInputAction.next,
       onChanged: (value) => onChange(value),
       scrollPhysics: const BouncingScrollPhysics(),
-      cursorColor: colorPrimary,
+      cursorColor: Theme.of(context).colorScheme.primary,
       validator: (value) => validator!(value),
       keyboardType: keyboardType == "number" ? TextInputType.number : TextInputType.text,
       style: Theme.of(context).textTheme.labelSmall,
