@@ -22,39 +22,39 @@ class NumberPadWidget extends StatelessWidget {
               crossAxisAlignment: CrossAxisAlignment.stretch,
               children: [
                 ButtonWidget(action: () => update("1"), label: "1"),
-                const SizedBox(height: mainUnit / 2),
+                SizedBox(height: Units().spacing / 2),
                 ButtonWidget(action: () => update("4"), label: "4"),
-                const SizedBox(height: mainUnit / 2),
+                SizedBox(height: Units().spacing / 2),
                 ButtonWidget(action: () => update("7"), label: "7"),
               ],
             ),
           ),
-          const SizedBox(width: mainUnit / 2),
+          SizedBox(width: Units().spacing / 2),
           Expanded(
             child: Column(
               mainAxisAlignment: MainAxisAlignment.start,
               crossAxisAlignment: CrossAxisAlignment.stretch,
               children: [
                 ButtonWidget(action: () => update("2"), label: "2"),
-                const SizedBox(height: mainUnit / 2),
+                SizedBox(height: Units().spacing / 2),
                 ButtonWidget(action: () => update("5"), label: "5"),
-                const SizedBox(height: mainUnit / 2),
+                SizedBox(height: Units().spacing / 2),
                 ButtonWidget(action: () => update("8"), label: "8"),
-                const SizedBox(height: mainUnit / 2),
+                SizedBox(height: Units().spacing / 2),
                 ButtonWidget(action: () => update("0"), label: "0"),
               ],
             ),
           ),
-          const SizedBox(width: mainUnit / 2),
+          SizedBox(width: Units().spacing / 2),
           Expanded(
             child: Column(
               mainAxisAlignment: MainAxisAlignment.start,
               crossAxisAlignment: CrossAxisAlignment.stretch,
               children: [
                 ButtonWidget(action: () => update("3"), label: "3"),
-                const SizedBox(height: mainUnit / 2),
+                SizedBox(height: Units().spacing / 2),
                 ButtonWidget(action: () => update("6"), label: "6"),
-                const SizedBox(height: mainUnit / 2),
+                SizedBox(height: Units().spacing / 2),
                 ButtonWidget(action: () => update("9"), label: "9"),
               ],
             ),
