@@ -1,5 +1,5 @@
-import 'package:lexa/view_model/theme/constants.dart';
 import 'package:flutter/material.dart';
+import 'package:lexa/view_model/theme/constants.dart';
 
 class InfoRecordWidget extends StatelessWidget {
   final String title;
@@ -15,10 +15,10 @@ class InfoRecordWidget extends StatelessWidget {
   Widget build(BuildContext context) {
     return Expanded(
       child: Container(
-        padding: const EdgeInsets.all(mainUnit / 2),
+        padding: EdgeInsets.all(Units().spacing / 2),
         decoration: BoxDecoration(
-          color: colorBackground,
-          borderRadius: BorderRadius.circular(borderRadius),
+          color: Theme.of(context).colorScheme.background,
+          borderRadius: BorderRadius.circular(Units().borderRadius),
         ),
         clipBehavior: Clip.hardEdge,
         child: Column(
@@ -26,8 +26,8 @@ class InfoRecordWidget extends StatelessWidget {
           crossAxisAlignment: CrossAxisAlignment.center,
           children: [
             FittedBox(child: Text(title, style: Theme.of(context).textTheme.titleMedium)),
-            const SizedBox(height: mainUnit / 6),
-            Text("${value.toString()} E£", style: Theme.of(context).textTheme.bodySmall?.copyWith(color: colorTertiary)),
+            SizedBox(height: Units().spacing / 6),
+            Text("${value.toString()} E£", style: Theme.of(context).textTheme.bodySmall?.copyWith(color: Theme.of(context).colorScheme.tertiary)),
           ],
         ),
       ),
